refactor (gravsearch): make new package sparql
