+++ webapi/src/main/scala/org/knora/webapi/responders/v2/SearchResponderV2.scala
@@ -30,8 +30,8 @@ import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util._
 import org.knora.webapi.util.search.ApacheLuceneSupport.{CombineSearchTerms, MatchStringWhileTyping}
 import org.knora.webapi.util.search._
-import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.SparqlTransformation._
-import org.knora.webapi.util.search.gravsearch._
+import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.SparqlTransformation._
+import org.knora.webapi.util.search.sparql.gravsearch._
 
 import scala.concurrent.Future
 
@@ -128,8 +128,8 @@ class SearchResponderV2 extends ResponderWithStandoffV2 {
       */
     private def fulltextSearchV2(searchValue: String, offset: Int, limitToProject: Option[IRI], limitToResourceClass: Option[SmartIri], limitToStandoffClass: Option[SmartIri], requestingUser: UserADM): Future[ReadResourcesSequenceV2] = {
 
-        import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.FulltextSearch._
-        import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.FulltextSearch.FullTextSearchConstants._
+        import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.FulltextSearch._
+        import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.FulltextSearch.FullTextSearchConstants._
 
         val groupConcatSeparator = StringFormatter.INFORMATION_SEPARATOR_ONE
 
@@ -360,7 +360,7 @@ class SearchResponderV2 extends ResponderWithStandoffV2 {
       */
     private def gravsearchV2(inputQuery: ConstructQuery, requestingUser: UserADM): Future[ReadResourcesSequenceV2] = {
 
-        import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.Gravsearch._
+        import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.Gravsearch._
 
         for {
             // Do type inspection and remove type annotations from the WHERE clause.
