+++ webapi/src/main/scala/org/knora/webapi/routing/v2/SearchRouteV2.scala
@@ -29,7 +29,7 @@ import org.knora.webapi.messages.v2.responder.searchmessages._
 import org.knora.webapi.responders.RESPONDER_MANAGER_ACTOR_PATH
 import org.knora.webapi.routing.{Authenticator, RouteUtilV2}
 import org.knora.webapi.util.IriConversions._
-import org.knora.webapi.util.search.gravsearch.GravsearchParser
+import org.knora.webapi.util.search.sparql.gravsearch.GravsearchParser
 import org.knora.webapi.util.{SmartIri, StringFormatter}
 
 import scala.concurrent.{ExecutionContext, Future}
