+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/gravsearch/AbstractSparqlTransformer.scala
@@ -1,12 +1,12 @@
-package org.knora.webapi.util.search.gravsearch
+package org.knora.webapi.util.search.sparql.gravsearch
 
 import org.knora.webapi._
 import org.knora.webapi.messages.v2.responder.valuemessages.DateValueContentV2
 import org.knora.webapi.util.{SmartIri, StringFormatter}
 import org.knora.webapi.util.search.ApacheLuceneSupport.CombineSearchTerms
 import org.knora.webapi.util.search._
-import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.Gravsearch.GravsearchConstants
-import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.SparqlTransformation._
+import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.Gravsearch.GravsearchConstants
+import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.SparqlTransformation._
 import org.knora.webapi.util.IriConversions._
 
 import scala.collection.mutable
