+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/gravsearch/GravsearchTypeInspectionRunner.scala
@@ -17,7 +17,7 @@
  * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.knora.webapi.util.search.gravsearch
+package org.knora.webapi.util.search.sparql.gravsearch
 
 import akka.actor.ActorSystem
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
