+++ webapi/src/test/scala/org/knora/webapi/util/search/sparql/gravsearch/GravsearchParserSpec.scala
@@ -17,7 +17,7 @@
  * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.knora.webapi.util.search.gravsearch
+package org.knora.webapi.util.search.sparql.gravsearch
 
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util.StringFormatter
